+++ plugins/O-SimpleReverb/Source/PluginEditor.h
@@ -13,9 +13,15 @@
 #pragma once
 #include "PluginProcessor.h"
 #include <juce_gui_extra/juce_gui_extra.h>
+#if OUARICON_LICENSING_ENABLED
+  #include "OuariconLicenseUI.h"
+#endif
 
 class OSimpleReverbAudioProcessorEditor : public juce::AudioProcessorEditor,
                                                   private juce::Timer
+#if OUARICON_LICENSING_ENABLED
+                                                , private OuariconLicense::Listener
+#endif
 {
 public:
     explicit OSimpleReverbAudioProcessorEditor(OSimpleReverbAudioProcessor&);
@@ -58,6 +64,11 @@ private:
 
     // File chooser for save/load dialogs
     std::unique_ptr<juce::FileChooser> fileChooser;
+
+#if OUARICON_LICENSING_ENABLED
+    std::unique_ptr<OuariconLicenseOverlay> licenseOverlay;
+    void licenseStatusChanged(OuariconLicense&, OuariconLicense::Status) override;
+#endif
 
     JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(OSimpleReverbAudioProcessorEditor)
 };